refactor hardware aggregation pipeline
